+++ src/methods/is-logged.js
@@ -1,3 +1,0 @@
-export default (self, session) => {
-  return Date.now() < new Date(session.expires).getTime() && Boolean(session.access_token)
-}