+++ src/methods/logout.js
@@ -1,9 +1,20 @@
-import emitter from './../lib/emitter'
+import emitter from '../lib/emitter'
 
-export default (self, session) => {
-  // just clears current session
-  session.access_token = ''
-  // emit logout event
+/**
+ * @method
+ * @name EcomAuth#logout
+ * @description Reset admin session object.
+ *
+ * @returns {self}
+ *
+ * @example
+
+ecomAuth.logout()
+
+ */
+
+export default (self) => {
+  self.session.access_token = ''
   emitter.emit('logout', self)
   return self
 }